BTA-410 | renamed tests and added to help and contact controller
