+++ app/controllers/HelpAndContactController.scala
@@ -17,7 +17,6 @@
 package controllers
 
 import javax.inject.Inject
-
 import config.FrontendAppConfig
 import controllers.actions._
 import handlers.ErrorHandler
@@ -28,6 +27,7 @@ import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.AnyContent
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
 import views.html.sa._
+import views.html.selfAssessment.self_assessment_tax_return_check
 import views.html.vat.{payments_and_deadlines, questions_about_vat}
 
 class HelpAndContactController @Inject()(appConfig: FrontendAppConfig,
@@ -55,6 +55,7 @@ class HelpAndContactController @Inject()(appConfig: FrontendAppConfig,
   private def selfAssessment(page: String)(implicit request: ServiceInfoRequest[AnyContent]) = {
     page match {
       case "how-to-pay" => Ok(how_to_pay_self_assessment(appConfig)(request.serviceInfoContent))
+      case "help-with-return" => Ok(self_assessment_tax_return_check(appConfig)(request.serviceInfoContent))
       case _ => NotFound(errorHandler.notFoundTemplate)
     }
   }
